# swagger.yaml
...
openapi: '3.1.0'
...
info:
  version: '0.0.1'
  title: Todo List API
servers:
  - url: http://localhost:4000/api
    description: Production server
schemes: -https
...
paths:
  /auth/login:
    post:
      tags:
        - auth
      description: Login api
      produces:
        - application/json
      requestBody:
        description: Login
        content:
          application/json:
            schema:
              $ref: '#/components/schemes/login'
      responses:
        200:
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/loginSuccess'
        401:
          description: Username or password is incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/401-isCorrect'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
  /auth/register:
    post:
      tags:
        - auth
      description: Register a account
      produces:
        - application/json
      requestBody:
        description: Register
        content:
          application/json:
            schema:
              $ref: '#/components/schemes/register'
      responses:
        200:
          description: Register success
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/registerSuccess'
        401:
          description: Username exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/401-userExists'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
  /auth/refresh-token:
    get:
      tags:
        - auth
      description: Get refresh token
      parameters:
        - in: header
          name: id
          schema:
            type: integer
          required: true
          description: id user account
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/refreshTokenSuccess'
        401:
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/401-invalidRefreshToken'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
  /auth/reset-password:
    post:
      tags:
        - auth
      description: Reset password
      produces:
        - application/json
      requestBody:
        description: Reset password
        content:
          application/json:
            schema:
              $ref: '#/components/schemes/reset-password'
      responses:
        200:
          description: Reset password successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/resetPasswordSuccess'
        400:
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/400-invalidOTP'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
  /auth/send-mail-reset:
    post:
      tags:
        - auth
      description: send mail reset
      produces:
        - application/json
      requestBody:
        description: send mail reset
        content:
          application/json:
            schema:
              $ref: '#/components/schemes/send-mail-reset'
      responses:
        200:
          description: Send email successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/sendMailResetSuccess'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
  /auth/change-password:
    post:
      tags:
        - auth
      description: change password
      produces:
        - application/json
      security:
        - apiKey: []
      requestBody:
        description: change password
        content:
          application/json:
            schema:
              $ref: '#/components/schemes/change-password'
      responses:
        200:
          description: Send email successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/changePasswordSuccess'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
...
  /account:
    get:
      tags:
        - account
      description: get all accounts, requires api key
      produces:
        - application/json
      security:
        - apiKey: []
      responses:
        200:
          description: Get accounts success
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/getAllAccountsSuccess'
        401:
          description: JWT expired or invalid
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/401-JWT-expired'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
    post:
      tags:
        - account
      description: add new account, requires api key
      produces:
        - application/json
      security:
        - apiKey: []
      requestBody:
        description: add new account
        content:
          application/json:
            schema:
              $ref: '#/components/schemes/add-account'
      responses:
        201:
          description: Add account success
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/addAccountSuccess'
        401:
          description: JWT expired or invalid
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/401-JWT-expired'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
  /account/{id}:
    get:
      tags:
        - account
      description: find one account by id, requires api key
      produces:
        - application/json
      security:
        - apiKey: []
      parameters:
        - name: id
          in: path
          description: id
          required: true
          explode: true
          schema:
            type: number
      responses:
        200:
          description: Get account success
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/getAccountSuccess'
        400:
          description: Invalid id, requires format numeric
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/400-invalid-format-id'
        401:
          description: JWT expired or invalid
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/401-JWT-expired'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
    put:
      tags:
        - account
      description: update account, requires api key
      produces:
        - application/json
      security:
        - apiKey: []
      parameters:
        - name: id
          in: path
          description: id
          required: true
          explode: true
          schema:
            type: string
      requestBody:
        description: update account
        content:
          application/json:
            schema:
              $ref: '#/components/schemes/add-account'
      responses:
        200:
          description: update account success
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/updateAccountSuccess'
        400:
          description: Invalid id, requires format numeric
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/400-invalid-format-id'
        401:
          description: JWT expired or invalid
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/401-JWT-expired'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
    delete:
      tags:
        - account
      description: delete account by id, requires api key
      produces:
        - application/json
      security:
        - apiKey: []
      parameters:
        - name: id
          in: path
          description: id
          required: true
          explode: true
          schema:
            type: string
      responses:
        200:
          description: delete account success
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/deleteAccountSuccess'
        400:
          description: Invalid id, requires format numeric
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/400-invalid-format-id'
        401:
          description: JWT expired or invalid
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemes/401-JWT-expired'
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemes/500'
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      name: Authorization
      in: header
      format: JWT
  schemes:
    login:
      required:
        - username
        - password
      type: object
      properties:
        username:
          type: string
          description: Username of account
        password:
          type: string
          description: Password of account
      example:
        username: xuanlong0
        password: '12345'
    register:
      require:
        - profile
        - account
      type: object
      properties:
        profile:
          type: object
          properties:
            name:
              type: string
            dob:
              type: string
            gender:
              type: string
            email:
              type: string
            image:
              type: string
        account:
          type: object
          properties:
            username:
              type: string
            password:
              type: string
            role:
              type: string
      example:
        profile:
          name: long
          dob: 2002-11-19
          gender: male
          email: long@gmail.com
          image: https://static.tvtropes.org/pmwiki/pub/images/2e5c6d37_566f_4274_b62d_ebf5fcbd0722.png
        account:
          username: a1231a
          password: Admin123
...
    reset-password:
      require:
        - username
        - password
        - otp
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        otp:
          type: string
      example:
        username: xuanlong0
        password: '12345'
        otp: '191212'
    send-mail-reset:
      require:
        - email
      type: object
      properties:
        email:
          type: string
      example:
        email: 'hello@gmail.com'
    change-password:
      require:
        - password
      type: object
      properties:
        password:
          type: string
      example:
        password: '12345'
...
    add-account:
      type: object
      properties:
        id:
          type: number
        username:
          type: string
        password:
          type: string
        role:
          type: string
        profile:
          type: object
          properties:
            id:
              type: number
            name:
              type: string
            dob:
              type: string
            gender:
              type: string
            email:
              type: string
            image:
              type: string
      example:
        username: xuanlong1123
        password: $2b$05$HWa7ZP7IHEHRp9hm.5tgZ.k/9dNIxC/csTpLHDLrJa.C6xE4.2jDS
        role: member
        profile:
          id: 4
          name: long
          dob: 2002-11-19
          gender: male
          email: long@gmail.com
          image: https://static.tvtropes.org/pmwiki/pub/images/2e5c6d37_566f_4274_b62d_ebf5fcbd0722.png
...
    loginSuccess:
      type: object
      properties:
        message:
          type: string
        access_token:
          type: string
        refresh_token:
          type: string
        data:
          type: object
          properties:
            id:
              type: number
            username:
              type: string
            role:
              type: string
            provider:
              type: object
              properties:
                id:
                  type: number
                name:
                  type: string
                dob:
                  type: string
                gender:
                  type: string
                email:
                  type: string
                password:
                  type: string
      example:
        message: Login successful
        access_token: gdadadjadgagdw.adhagd.adgww
        refresh_token: gadauwd.hehdadw.awbdwha
        data:
          id: 3
          username: xuanlong0
          role: member
          profile:
            id: 3
            name: long
            dob: 2002-11-19
            gender: male
            email: long@gmail.com
            image: jdadhaudwgdwyadg
    registerSuccess:
      type: object
      properties:
        message:
          type: string
        data:
          profile:
            type: object
            properties:
              name:
                type: string
              dob:
                type: string
              gender:
                type: string
              email:
                type: string
              image:
                type: string
          account:
          type: object
          properties:
            username:
              type: string
            password:
              type: string
            role:
              type: string
      example:
        message: Register successful
        data:
          username: a1231a
          password: $2b$10$7dcLwHCWd6weMBkr6lWcQuNOD8EoGCiS0APKKY/ZTlC8/yh4gVR5u
          role: member
          profile:
            name: long
            dob: 2002-11-19
            gender: male
            email: long@gmail.com
            image: https://static.tvtropes.org/pmwiki/pub/images/2e5c6d37_566f_4274_b62d_ebf5fcbd0722.png
            id: 10
          id: 12
    refreshTokenSuccess:
      type: object
      properties:
        message:
          type: string
        data:
          type: object
          properties:
            id:
              type: number
            username:
              type: string
            password:
              type: string
            role:
              type: string
            iat:
              type: string
        access_token:
          type: string
        refresh_token:
          type: string
      example:
        message: Success
        data:
          id: 5
          username: long
          password: 12313
          role: member
          iat: 18192371
        access_token: jdajdaahw.auhdw.12813bda
        refresh_token: ahdadhaw.2812.dgad
    resetPasswordSuccess:
      type: object
      properties:
        message:
          type: string
      example:
        message: Success
    sendMailResetSuccess:
      type: object
      properties:
        message:
          type: string
      example:
        message: Success
    changePasswordSuccess:
      type: object
      properties:
        message:
          type: string
      example:
        message: Success
...
    getAllAccountsSuccess:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: number
              username:
                type: string
              password:
                type: string
              role:
                type: string
              profile:
                type: object
                properties:
                  id:
                    type: number
                  name:
                    type: string
                  dob:
                    type: string
                  gender:
                    type: string
                  email:
                    type: string
                  image:
                    type: string
      example:
        data:
          - id: 4
            username: xuanlong1123
            password: $2b$05$HWa7ZP7IHEHRp9hm.5tgZ.k/9dNIxC/csTpLHDLrJa.C6xE4.2jDS
            role: member
            profile:
              id: 4
              name: long
              dob: 2002-11-19
              gender: male
              email: long@gmail.com
              image: https://static.tvtropes.org/pmwiki/pub/images/2e5c6d37_566f_4274_b62d_ebf5fcbd0722.png
    addAccountSuccess:
      type: object
      properties:
        id:
          type: number
        username:
          type: string
        password:
          type: string
        role:
          type: string
        profile:
          type: object
          properties:
            id:
              type: number
            name:
              type: string
            dob:
              type: string
            gender:
              type: string
            email:
              type: string
            image:
              type: string
      example:
        message: 'Account id 4 created'
        data:
          id: 4
          username: xuanlong1123
          password: $2b$05$HWa7ZP7IHEHRp9hm.5tgZ.k/9dNIxC/csTpLHDLrJa.C6xE4.2jDS
          role: member
          profile:
            id: 4
            name: long
            dob: 2002-11-19
            gender: male
            email: long@gmail.com
            image: https://static.tvtropes.org/pmwiki/pub/images/2e5c6d37_566f_4274_b62d_ebf5fcbd0722.png
    getAccountSuccess:
      type: object
      properties:
        id:
          type: number
        username:
          type: string
        password:
          type: string
        role:
          type: string
        profile:
          type: object
          properties:
            id:
              type: number
            name:
              type: string
            dob:
              type: string
            gender:
              type: string
            email:
              type: string
            image:
              type: string
      example:
        data:
          id: 4
          username: xuanlong1123
          password: $2b$05$HWa7ZP7IHEHRp9hm.5tgZ.k/9dNIxC/csTpLHDLrJa.C6xE4.2jDS
          role: member
          profile:
            id: 4
            name: long
            dob: 2002-11-19
            gender: male
            email: long@gmail.com
            image: https://static.tvtropes.org/pmwiki/pub/images/2e5c6d37_566f_4274_b62d_ebf5fcbd0722.png
    updateAccountSuccess:
      type: object
      properties:
        id:
          type: number
        username:
          type: string
        password:
          type: string
        role:
          type: string
        profile:
          type: object
          properties:
            id:
              type: number
            name:
              type: string
            dob:
              type: string
            gender:
              type: string
            email:
              type: string
            image:
              type: string
      example:
        data:
          id: 4
          username: xuanlong1123
          password: $2b$05$HWa7ZP7IHEHRp9hm.5tgZ.k/9dNIxC/csTpLHDLrJa.C6xE4.2jDS
          role: member
          profile:
            id: 4
            name: long
            dob: 2002-11-19
            gender: male
            email: long@gmail.com
            image: https://static.tvtropes.org/pmwiki/pub/images/2e5c6d37_566f_4274_b62d_ebf5fcbd0722.png
    deleteAccountSuccess:
      type: object
      properties:
        message:
          type: string
      example:
        message: success
...
    400-invalidOTP:
      type: object
      properties:
        error:
          type: string
      example:
        error: Invalid OTP
    400-invalid-format-id:
      type: object
      properties:
        error:
          type: string
      example:
        error: Validation failed (numeric string is expected)
...
    401-isCorrect:
      type: object
      properties:
        error:
          type: string
      example:
        error: Username or password is incorrect
    401-userExists:
      type: object
      properties:
        error:
          type: string
      example:
        error: Username exists
    401-JWT-expired:
      type: object
      properties:
        error:
          type: string
      example:
        error: Invalid or expired JWT token
    401-invalidRefreshToken:
      type: object
      properties:
        error:
          type: string
      example:
        error: Invalid refresh token
    500:
      type: object
      properties:
        error:
          type: string
      example:
        error: Internal server error
# security:
#   - apiKey: []